+++ src/number_theory/extended_euclidean.cpp
@@ -5,8 +5,8 @@
     ax + by = 1を満たす整数x,yを求める
 */
 
-int extgcd(int a, int b, int &x, int &y) {
-  int d = a;
+long long int extgcd(long long int a, long long int b, long long int &x, long long int &y) {
+  long long int d = a;
   if (b != 0) {
     d = extgcd(b, a % b, y, x);
     y -= (a / b) * x;
