+++ src/number_theory/sieve.cpp
@@ -12,19 +12,19 @@
 
 #define MAX_N 10000
 
-int prime[MAX_N];         // i番目の素数
+long long int prime[MAX_N];         // i番目の素数
 bool is_prime[MAX_N + 1]; // is_prime[i]がtrueならばiは素数
 
 // n以下の素数の数を返す
-int sieve(int n) {
-  int p = 0;
-  for (int i = 0; i <= n; i++)
+long long int sieve(long long int n) {
+  long long int p = 0;
+  for (long long int i = 0; i <= n; i++)
     is_prime[i] = true;
   is_prime[0] = false;
-  for (int i = 2; i <= n; i++) {
+  for (long long int i = 2; i <= n; i++) {
     if (is_prime[i]) {
       prime[p++] = i;
-      for (int j = 2 * i; j <= n; j += i)
+      for (long long int j = 2 * i; j <= n; j += i)
         is_prime[j] = false;
     }
   }
